pyproject.toml
[tool.poetry]
name = "fast-chat"
version = "0.1.0"
description = ""
authors = ["kk <kanbang@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7, <3.8"
pytz = "^2024.2"
fastapi = "0.103.2"
chatterbot = "^1.0.8"
uvicorn = "0.22.0"
numpy = "1.21.6"
spacy = "2.3.5"
jinja2 = "^3.1.4"
pyyaml = "5.4"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
